roles

+++ src/Controller/UserProfilController.php
@@ -40,49 +40,56 @@ class UserProfilController extends AbstractController
      */
     public function updateProfil(UsersRepository $usersRepository, $id, Request $request, UserPasswordEncoderInterface $passwordEncoder, SkillsRepository $skillsRepository ){
         $user = $usersRepository->find($id);
+        $userlogin = $this->getUser();
         // $skills = $skillsRepository->findAll();
-        $form = $this->createForm(ProfilUserType::class, $user);
-        $form->handleRequest($request);
+        if($userlogin == $user){
 
-        if($form->isSubmitted() && $form->isValid())
-        {
-            if($form->get('plainPassword')->getData() !== null) 
+            $form = $this->createForm(ProfilUserType::class, $user);
+            $form->handleRequest($request);
+    
+            if($form->isSubmitted() && $form->isValid())
             {
-                $user->setPassword(
-                    $passwordEncoder->encodePassword(
-                        $user,
-                        $form->get('plainPassword')->getData()
-                    )
-                );
-            }
-
-            // image profil
-
-            
-            $infoImg1 = $form['imgprofil']->getData();
-            $oldNomImg1 = $user->getImgprofil();
-
-            if($infoImg1!=null){
-                $oldCheminImg1 = $this->getParameter('photos_users') . '/' . $oldNomImg1;       
-                if (file_exists($oldCheminImg1)) 
+                if($form->get('plainPassword')->getData() !== null) 
                 {
-                    unlink($oldCheminImg1);
+                    $user->setPassword(
+                        $passwordEncoder->encodePassword(
+                            $user,
+                            $form->get('plainPassword')->getData()
+                        )
+                    );
                 }
-                $extensionImg1 = $infoImg1->guessExtension();
-                $nomImg1 = '1-' . time() . '.' . $extensionImg1;
-                $infoImg1->move($this->getParameter('photos_users'), $nomImg1);
-                $user->setImgprofil($nomImg1);
+    
+                // image profil
+    
+                
+                $infoImg1 = $form['imgprofil']->getData();
+                $oldNomImg1 = $user->getImgprofil();
+    
+                if($infoImg1!=null){
+                    $oldCheminImg1 = $this->getParameter('photos_users') . '/' . $oldNomImg1;       
+                    if (file_exists($oldCheminImg1)) 
+                    {
+                        unlink($oldCheminImg1);
+                    }
+                    $extensionImg1 = $infoImg1->guessExtension();
+                    $nomImg1 = '1-' . time() . '.' . $extensionImg1;
+                    $infoImg1->move($this->getParameter('photos_users'), $nomImg1);
+                    $user->setImgprofil($nomImg1);
+                }
+                
+    
+                $manager = $this->getDoctrine()->getManager();
+                $manager->persist($user);
+                $manager->flush();
+                // redirection vers profil user courant 
+                return $this->redirectToRoute('user_profil', ['id' => $id]);
             }
-            
-
-            $manager = $this->getDoctrine()->getManager();
-            $manager->persist($user);
-            $manager->flush();
-            // redirection vers profil user courant 
-            return $this->redirectToRoute('user_profil', ['id' => $id]);
+            return $this->render('user_profil/userProfilForm.html.twig', [
+                'userProfilForm' => $form->createView()
+            ]);
         }
-        return $this->render('user_profil/userProfilForm.html.twig', [
-            'userProfilForm' => $form->createView()
-        ]);
+        else {
+            return $this->render('home/home.html.twig');
+        }
     }
 }
